a new tests for create_scheduler_from_config were added. test_bad_scheduler and test_bad_scheduler_fields_file moved into test_scheduler
