+++ pocs/tests/test_scheduler.py
@@ -1,0 +1,32 @@
+import pytest
+
+from pocs.scheduler import create_scheduler_from_config
+from pocs.site_location_details import setup_site_location_details_from_config
+from pocs.utils import error
+
+
+def test_bad_scheduler_type(config):
+    conf = config.copy()
+    conf['scheduler']['type'] = 'foobar'
+    details = setup_site_location_details_from_config(config)
+    with pytest.raises(error.NotFound):
+        create_scheduler_from_config(conf, observer=details['observer'])
+
+
+def test_bad_scheduler_fields_file(config):
+    conf = config.copy()
+    conf['scheduler']['fields_file'] = 'foobar'
+    details = setup_site_location_details_from_config(config)
+    with pytest.raises(error.NotFound):
+        create_scheduler_from_config(conf, observer=details['observer'])
+
+
+def test_no_observer(config):
+    assert create_scheduler_from_config(config, observer=None) is None
+
+
+def test_no_scheduler_in_config(config):
+    conf = config.copy()
+    del conf['scheduler']
+    details = setup_site_location_details_from_config(conf)
+    assert create_scheduler_from_config(conf, observer=details['observer']) is None
